requirements.txt
langgraph
litellm
selenium
webdriver_manager
colorama
python-dotenv
flask
flask_cors
socksio
